+++ CHANGELOG.md
@@ -194,6 +194,7 @@ Listed changes are split into categories, reflecting the affected component / (s
 - Add created default archive trigger plugins to plugin overview in the docs
 - Add archiving trigger sub-plugin component and API descriptions to developer docs
 - Improve speed of artifact file SHA256 hash generation
+- Exclude deleted files from storage stats counters on components overview page
 
 
 ### Sub-Plugins